another signed/unsigned warning
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2757 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/hest/parseHest.c
@@ -683,10 +683,10 @@ _hestSetValues(char **prms, int *udflt, unsigned int *nprm, int *appr,
     _hestIdent(ident, opt+op, parm, AIR_TRUE);
     type = opt[op].type;
     size = (airTypeEnum == type
-            ? (int)sizeof(int)
+            ? (int)sizeof(int)             /* HEY scrutinize casts */
             : (airTypeOther == type
-               ? (int)opt[op].CB->size
-               : airTypeSize[type]));
+               ? (int)opt[op].CB->size     /* HEY scrutinize casts */
+               : (int)airTypeSize[type])); /* HEY scrutinize casts */
     cP = (char *)(vP = opt[op].valueP);
     if (parm->verbosity) {
       printf("%s %d of %d: \"%s\": |%s| --> kind=%d, type=%d, size=%d\n",
